refactor: BazelStats structure
