+++ src/main/java/com/fyp_poc/demo/DTO/BazelStats.java
@@ -2,9 +2,7 @@ package com.fyp_poc.demo.DTO;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
@@ -15,6 +13,8 @@ import java.util.UUID;
 @Getter
 @Setter
 @NoArgsConstructor
+@Builder
+@AllArgsConstructor
 @Entity(name="bazel_stats")
 @Table(name="bazel_stats")
 public class BazelStats {
