+++ src/main/java/com/fyp_poc/demo/functional/Aggregator.java
@@ -1,0 +1,6 @@
+package com.fyp_poc.demo.functional;
+
+@FunctionalInterface
+public interface Aggregator<X extends Number> {
+    X process(X arg1, X arg2);
+}
