+++ src/main/java/com/fyp_poc/demo/services/bazelStats/BazelStatsAggregatorService.java
@@ -1,0 +1,11 @@
+package com.fyp_poc.demo.services.bazelStats;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BazelStatsAggregatorService implements IBazelStatsAggregatorService{
+
+
+
+
+}
