+++ src/main/java/com/fyp_poc/demo/services/bazelStats/IBazelStatsAggregatorService.java
@@ -1,0 +1,4 @@
+package com.fyp_poc.demo.services.bazelStats;
+
+public interface IBazelStatsAggregatorService {
+}
